Implement `IssuesEvent`s
These events are for when issues are opened or closed.

This also implements a `truncateString` method. Discord has a strict
embed size limit of 6000 after adding all descriptions, field names and
titles, etc. for all embeds in a message, and understandably, the GitHub
webhook is conservative and truncates issue descriptions to 500
characters.

This also changes the case-switch in index.js to be a lambda, so
`message` can be const.